Extract MillisecondsToDate to static function, fix receiving CancellationDate from Apple
